test(stackable-versioned): Add integration test for merged_crd
